Move more files to geo package
