+++ sql/catalyst/src/main/java/org/apache/spark/sql/catalyst/util/geo/STUtils.java
@@ -14,10 +14,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.spark.sql.catalyst.util;
+package org.apache.spark.sql.catalyst.util.geo;
 
-import org.apache.spark.sql.catalyst.util.geo.Geography;
-import org.apache.spark.sql.catalyst.util.geo.Geometry;
 import org.apache.spark.sql.errors.QueryExecutionErrors;
 import org.apache.spark.sql.types.GeographyType;
 import org.apache.spark.sql.types.GeometryType;
