+++ sql/catalyst/src/test/scala/org/apache/spark/sql/catalyst/util/geo/StUtilsSuite.java
@@ -18,7 +18,6 @@
 package org.apache.spark.sql.catalyst.util.geo;
 
 import org.apache.spark.SparkIllegalArgumentException;
-import org.apache.spark.sql.catalyst.util.STUtils;
 import org.apache.spark.unsafe.types.GeographyVal;
 import org.apache.spark.unsafe.types.GeometryVal;
 import org.junit.jupiter.api.Test;
